Calculator

+++ calculator/src/main/java/com/pechuro/cashdebts/calculator/Calculator.kt
@@ -1,11 +1,13 @@
 package com.pechuro.cashdebts.calculator
 
+import java.math.BigDecimal
+
 interface Calculator {
 
     fun evaluate(expr: String): Result
 }
 
 sealed class Result {
-    class Success(val result: Double) : Result()
+    class Success(val result: BigDecimal) : Result()
     object Error : Result()
 }
